+++ unbrake-frontend/src/components/Calibration/CalibrationTempFields.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../App.css";
+import "../../App.css";
 import { Field } from "redux-form";
 import { TextField, withStyles } from "@material-ui/core";
 import PropTypes from "prop-types";
