+++ unbrake-frontend/src/components/Calibration/CalibrationTempForm.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../App.css";
+import "../../App.css";
 import { reduxForm } from "redux-form";
 import CalibrationFields from "./CalibrationTempFields";
 import CalibrationCheck from "./CalibrationTempCheck";
